requirements.txt
discord.py==2.4.0
asyncio==3.4.3
gTTS==2.4.0
